Ignore measures that define groups
This update stops the code from trying to split the group based on a
measure that defines the groups already. For example if a data frame has
been defined to be only boys then split on a wears_bra variable and then
split on gender this will (most probably) only give one group (of a
possible four).